+++ CHANGES.rst
@@ -2,3 +2,4 @@
 ------------------
 
 * base dependencies + extras (halotools, h5py); install all dependencies via pip nbodykit[extras]
+* meta-data calculations in FKPCatalog now account for Source selection properly